Extracts transformers into own files
