+++ lib/helpers/reports/transformers/HeardFrom.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const HeaderField = require('./HeaderField');
+
+module.exports = class HeardFrom extends HeaderField {
+  title() {
+    return 'Heard from';
+  }
+
+  value() {
+    return this.attachments.length + ' people';
+  }
+
+  shortValue() {
+    return true;
+  }
+};
