+++ lib/helpers/reports/transformers/index.js
@@ -1,0 +1,4 @@
+'use strict';
+
+const requireDirectory = require('require-directory');
+module.exports = requireDirectory(module);
