+++ package.json
@@ -32,6 +32,7 @@
     "pg": "^6.1.2",
     "pg-hstore": "^2.3.2",
     "request": "^2.69.0",
+    "require-directory": "^2.1.1",
     "sequelize": "^3.24.3",
     "sequelize-cli": "^2.3.1",
     "underscore": "^1.8.3"
